requirements.txt
filetype==1.2.0
Flask==3.0.2
flask_cors==4.0.0
librosa==0.10.1
noisereduce==3.0.2
openai==1.14.3
pyffmpeg==2.4.2.18.1
python-dotenv==1.0.1
scipy
soundfile
